py/runtime: Allow multiple *args in function call
This is a partial implementation of PEP 448 to allow unmacking multiple
star args in a function or method call.

This is implemented by changing the emitted bytecodes so that both
positoinal args and star args are stored as positoinal args. A bitmap
is added to indicate if an argument at a given position is a positoinal
argument or a star arg.

In the bytecodes, this new bitmap takes the place of the old star arg.
It is stored as a small int, so this means only the first N arguments
can be star args where N is the number of bits in a small int.

The runtime is modified to interpret this new bytecode format while
still trying to perform as few memory reallocations as possible.