github/workflows/code_size: Comment instead of fail.
This changes the code size workflow to post a comment on pull requests
with the code size report. It also removes the error threshold so that
the test won't fail if code size increases.

Allowable code size changes are subjective, so shouldn't cause CI to
fail. In addition, failing CI tests can cause other hooks like code
coverage reports to be suppressed, so this fixes that problem as well.

Fixes: https://github.com/micropython/micropython/issues/8464
Signed-off-by: David Lechner <david@pybricks.com>